Add WebAssembly target
This is accomplished by inverting the game loop to make it more like a
javascript generator function. There's now a single function `game_move()`
that takes in a line of user input and returns the prompt for the next input
line (this is only required for the Loud Room, which has a special prompt).

This was a surprisingly easy transformation to do on the code since it keeps
all state in global variables, and only uses local variables in passing.

However, there are several places in the game where it asks you interactive
yes/no questions. For now, all these just assume you wanted to type
"yes". Those are deep in a callchain and are more difficult to make them
"yield".

The "more" functionality is also turned off for the same reason.