dinit.c: Add dumping code when compiled with -DDUMP_DTEXT
The idea is to remove the encrypted file image and use the structures (all
the ints are encyrption enough!) but reorganize the text so that we can
format it for the web more nicely. This requires access to the unencrypted
text, so this dump decrypts it. The canonical text will live as an easily
editable text file and be encrypted and compiled into C at make time.