Add some hacky markup to dtextc.txt and parse it (in encode.rb) into text (cli) or html (wasm).
dungeo.js now expects html output and has some CSS for the classes that the
wasm emits. This allows the output to look good on phones which are skinnier
than 80 columns.

The text output isn't 1 to 1 from the original--some right justified and
centered lines are different.

I tested against the original dtext with:

    ./encode.rb dtextc.txt --format=text --dump | diff -u <(git show 6752adfd0d02e655b97bb9b02c008227e56dd025:dtextc.txt) -

and adjusted things until the diff was as small as it could get.